requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
osmnx==1.7.1
requests==2.31.0
pydantic==2.5.0
gunicorn==21.2.0
httpx==0.25.0
python-dotenv==1.0.0
